+++ packages/adapters/raycast/src/index.ts
@@ -115,3 +115,4 @@ export {
 export * from "./project-search";
 export * from "./project-list-order";
 export * from "./project-list-limits";
+export * from "./native-scanner";